+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
-| Version |   Date   |       Changelog |
-| ------- | :------: | --------------: |
-| 0.0.1   | 10-01-23 | Initial Release |
+## 0.0.2 - _18.02.2023_
+
+**Patched**
+
+- justforlaughs.ext.setmyURL should work
+- Only a single tab will show memes
+
+## 0.0.1 - _10.01.2023_
+
+**Release**
+
+- Initial Release